feat: add better debug mode config and override the regular logger config if debug mode is enabled
