+++ internal/cli/help.go
@@ -24,7 +24,7 @@ import (
 	"strings"
 
 	"github.com/reginald-project/reginald/internal/flags"
-	"github.com/reginald-project/reginald/internal/logging"
+	"github.com/reginald-project/reginald/internal/log"
 	"github.com/reginald-project/reginald/internal/plugin"
 	"github.com/reginald-project/reginald/internal/terminal"
 	"github.com/reginald-project/reginald/internal/text"
@@ -307,20 +307,22 @@ func printHelp(cmd *plugin.Command, flagSet *flags.FlagSet, store *plugin.Store)
 // or the flag in the arguments list. It prints the help message of the command
 // that was given before the flag.
 func runHelp(ctx context.Context, cmd *plugin.Command, store *plugin.Store) error {
-	logging.Trace(ctx, "running help", "cmd", cmd)
+	log.Trace(ctx, "running help", "cmd", cmd)
 
 	root := rootCommand(cmd)
 	flagSet := newFlagSet()
 
-	logging.Trace(ctx, "resolved root", "root", root)
+	log.Trace(ctx, "resolved root", "root", root)
 
 	var found *plugin.Command
 
 Loop:
 	for _, arg := range os.Args[1:] {
-		logging.Trace(ctx, "checking arg", "arg", arg)
+		log.Trace(ctx, "checking arg", "arg", arg)
 
 		if arg == "-h" || arg == "--help" {
+			log.Trace(ctx, "found help flag")
+
 			break
 		}
 
@@ -341,6 +343,8 @@ Loop:
 		}
 
 		if root == nil {
+			log.Trace(ctx, "no command found and root is nil")
+
 			continue
 		}
