+++ internal/log/logconfig/logconfig.go
@@ -1,0 +1,27 @@
+// Copyright 2025 The Reginald Authors
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+// Package logconfig defines the configuration options for the logger. It is
+// a separate pckage to avoid import cycles.
+package logconfig
+
+import "github.com/reginald-project/reginald-sdk-go/logs"
+
+// Config contains the configuration options for the logger.
+type Config struct {
+	Format  string     `mapstructure:"format"`                    // format of the logs, "json" or "text"
+	Output  string     `mapstructure:"output"`                    // destination of the logs
+	Level   logs.Level `mapstructure:"level"`                     // logging level
+	Enabled bool       `flag:"log,no-log" mapstructure:"enabled"` // whether logging is enabled
+}
